[frontend] Add app/lib/types.ts
